refactor: remove dupe tests and move others in test_power.py
Created test_individual_power.py for those specific tests, and moved
another test to its more appropraite home in test_cluster_power.py.
test_power.py should now focus on the unified facade in front of our
individual/cluster approaches.
